Move JournalLineReference to file where it's used
